requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.7.6
    # via flask-migrate
cffi==1.15.0
    # via gevent
click==8.0.3
    # via flask
colorama==0.4.4
    # via click
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-migrate
    #   flask-sqlalchemy
flask-cors==3.0.10
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
gevent==21.12.0
    # via -r requirements.in
greenlet==1.1.2
    # via
    #   gevent
    #   sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
mako==1.1.6
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
mysqlclient==2.1.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pycryptodomex==3.14.1
    # via -r requirements.in
pyjwt==2.3.0
    # via -r requirements.in
python-dotenv==0.19.2
    # via -r requirements.in
six==1.16.0
    # via flask-cors
sqlalchemy==1.4.31
    # via
    #   alembic
    #   flask-sqlalchemy
werkzeug==2.0.2
    # via flask
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
